Code to help solve 2021 day 24
This did an analysis of the program using really janky symbolic math, based on patterns that came out while doing this by hand. Spotted the only combination of conditionals that allowed Z = 0, and then solved what the min and max values of the digits could be on paper. I'm sure it could be done with code, but it didn't feel worth it after struggling on this one for a while. Really cool puzzle.